# swagger.yaml
swagger: "2.0"
info:
  description: "Banking Application API documentation called Banka. Banka is a light-weight core banking application that powers banking operations like account creation, customer deposit and withdrawals. This app is meant to support a single bank, where users can signup and create bank accounts online, but must visit the branch to withdraw or deposit money."
  version: "1.0.0"
  title: "Banka"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "kech3443@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/api/v1"
...
schemes:
- "https"
- "http"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
paths:
...
  /admin/setuser/{userid}:
    post:
      tags:
      - "Admin"
      summary: "Updates user type"
      description: "Sets user type to either admin, cashier, adminstaff or client "
      operationId: "updateUserType"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - Bearer: []
      parameters:
      - in: "body"
        name: "admin"
        description: "usertype object that is required to update the database"
        required: true
        schema:
          $ref: "#/definitions/updateUserType"
      - in: "path"
        name: "userid"
        description: "userid parameter that is required to update the database"
        required: true
        schema:
          $ref: "#/definitions/updateUserType"
      responses:
        200:
          description: "Successful Operation"
        400:
          description: "Bad Request"
        404:
          description: "User ID not Found"
...
definitions:
...
  updateUserType:
    type: "object"
    required:
      - usertype
    properties:
      usertype:
        type: "string"
        uniqueItems: false
...